+++ electron/config.js
@@ -54,6 +54,7 @@ const schema = {
             "tabSize": {type: "integer", default: 4},
             "defaultBlockLanguage": {type: "string"},
             "defaultBlockLanguageAutoDetect": {type: "boolean"},
+            "spellcheckEnabled": {type: "boolean", default:false},
 
             // when default font settings are used, fontFamily and fontSize is not specified in the 
             // settings file, so that it's possible for us to change the default settings in the 
@@ -127,6 +128,7 @@ const defaults = {
             wholeWord: false,
             regexp: false,
         },
+        spellcheckEnabled: false,
     },
     theme: "system",
 }