+++ docs/changelog.md
@@ -19,6 +19,7 @@ It's now possible to turn on/off spellchecking. This can be done by clicking the
 - Remember the open buffer(s) between program restarts.
 - Fixed so that Alt-clicking on an cursor removes it (if more than one cursor exists).
 - Changed the key binding for rectangular selection from Alt+Click to Alt+Shift+Click.
+- Fixed issue where the window position wouldn't be properly restored in some cases.
 - Updated to latest version of Electron.